+++ e2-studio-star-rx72n-firmware/libs/rx_hcsr04/src/rx_hcsr04_hal_hw.c
@@ -1048,7 +1048,7 @@ static void internal_cmt2_init(void)
   cmstr1 = &(cmt_ctrl()->cmstr1);
 
   /* Stop CMT2 (bit 0 of CMSTR1) */
-  *cmstr1 &= ~k_cmstr1_cmt2_enable_bit;
+  *cmstr1 &= (uint16_t)~(uint16_t)k_cmstr1_cmt2_enable_bit;
 
   /* Configure CMT2: PCLK/8, no interrupt */
   cmt2()->cmcr = k_cmt2_divider_bits;
